src/main/java/com/fiany/wechat/service/impl/ProductCategoryServiceImpl.java
package com.fiany.wechat.service.impl;

import com.fiany.wechat.dataobject.ProductCategory;
import com.fiany.wechat.repository.ProductCategoryRepository;
import com.fiany.wechat.service.IProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @Description : 类目
 * @Author : yifan
 * @Data : 2018/3/22 22:07
 */
@Service
public class ProductCategoryServiceImpl implements IProductCategoryService {
    @Autowired
    private ProductCategoryRepository repository;
    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findById(categoryId).get();
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
